requirements.txt
# Core Data Engineering
pyspark==3.4.1
psycopg2-binary==2.9.9
pandas==2.0.3
pyyaml==6.0.1

# dbt for Analytics
dbt-core==1.8.7
dbt-postgres==1.8.2

# Data Quality
great-expectations==0.18.19

# Utilities
python-dotenv==1.0.0